# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.CreateItemVendaDTO:
    properties:
      produto_id:
        type: string
      quantidade:
        type: integer
    required:
    - produto_id
    - quantidade
    type: object
  domain.CreateVendaDTO:
    properties:
      cliente:
        type: string
      desconto:
        maximum: 100
        minimum: 0
        type: number
      itens:
        items:
          $ref: '#/definitions/domain.CreateItemVendaDTO'
        type: array
    required:
    - cliente
    - itens
    type: object
  domain.ItemVenda:
    properties:
      id:
        type: string
      preco_unitario:
        type: number
      produto:
        $ref: '#/definitions/domain.Produto'
      produto_id:
        type: string
      quantidade:
        type: integer
      venda_id:
        type: string
    type: object
  domain.Produto:
    properties:
      data_criacao:
        type: string
      descricao:
        type: string
      id:
        type: string
      imagem_url:
        type: string
      nome:
        type: string
      preco:
        type: number
      quantidade:
        type: integer
    type: object
  domain.Role:
    enum:
    - admin
    - vendedor
    - cliente
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleVendedor
    - RoleCliente
  domain.Usuario:
    properties:
      ativo:
        type: boolean
      data_criacao:
        type: string
      email:
        type: string
      id:
        type: string
      nome:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
    type: object
  domain.Venda:
    properties:
      cliente:
        $ref: '#/definitions/domain.Usuario'
      cliente_id:
        type: string
      data_criacao:
        type: string
      data_venda:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.ItemVenda'
        type: array
      valor_total:
        type: number
      vendedor:
        $ref: '#/definitions/domain.Usuario'
      vendedor_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: pedro@example.com
    name: Pedro
  description: API para gerenciamento de vendas e produtos
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Sistema de Vendas API
  version: "1.0"
paths:
  /produtos:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de todos os produtos cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Produto'
            type: array
      summary: Lista todos os produtos
      tags:
      - produtos
    post:
      consumes:
      - application/json
      description: Cria um novo produto com os dados fornecidos
      parameters:
      - description: Dados do produto
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/domain.Produto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Produto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria um novo produto
      tags:
      - produtos
  /produtos/{id}:
    delete:
      consumes:
      - application/json
      description: Remove um produto pelo seu ID
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove um produto
      tags:
      - produtos
    get:
      consumes:
      - application/json
      description: Retorna um produto específico pelo seu ID
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Produto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtém um produto por ID
      tags:
      - produtos
    put:
      consumes:
      - application/json
      description: Atualiza um produto existente com os dados fornecidos
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      - description: Dados do produto
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/domain.Produto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Produto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza um produto
      tags:
      - produtos
  /relatorios:
    get:
      consumes:
      - application/json
      description: Retorna dados gerais do sistema como vendas do dia, total de clientes
        e produtos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Obtém relatório geral
      tags:
      - relatorios
  /vendas:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de todas as vendas cadastradas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Venda'
            type: array
      summary: Lista todas as vendas
      tags:
      - vendas
    post:
      consumes:
      - application/json
      description: Cria uma nova venda com os dados fornecidos
      parameters:
      - description: Dados da venda
        in: body
        name: venda
        required: true
        schema:
          $ref: '#/definitions/domain.CreateVendaDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Venda'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria uma nova venda
      tags:
      - vendas
  /vendas/{id}:
    delete:
      consumes:
      - application/json
      description: Remove uma venda pelo seu ID
      parameters:
      - description: ID da venda
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove uma venda
      tags:
      - vendas
    get:
      consumes:
      - application/json
      description: Retorna uma venda específica pelo seu ID
      parameters:
      - description: ID da venda
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Venda'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtém uma venda por ID
      tags:
      - vendas
    put:
      consumes:
      - application/json
      description: Atualiza uma venda existente com os dados fornecidos
      parameters:
      - description: ID da venda
        in: path
        name: id
        required: true
        type: string
      - description: Dados da venda
        in: body
        name: venda
        required: true
        schema:
          $ref: '#/definitions/domain.Venda'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Venda'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza uma venda
      tags:
      - vendas
  /vendas/cliente/{cliente}:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de vendas filtrada por cliente
      parameters:
      - description: Nome do cliente
        in: path
        name: cliente
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Venda'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista vendas por cliente
      tags:
      - vendas
  /vendas/periodo/{inicio}/{fim}:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de vendas filtrada por período
      parameters:
      - description: Data inicial (timestamp)
        in: path
        name: inicio
        required: true
        type: integer
      - description: Data final (timestamp)
        in: path
        name: fim
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Venda'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista vendas por período
      tags:
      - vendas
schemes:
- http
swagger: "2.0"
tags:
- description: Operações relacionadas a produtos
  name: produtos
- description: Operações relacionadas a vendas
  name: vendas